Separate getCalendarMonthGridWidthStyle to its own util file
