+++ src/components/CalendarMonthGrid.jsx
@@ -10,7 +10,7 @@ import CalendarMonth from './CalendarMonth';
 import isTransitionEndSupported from '../utils/isTransitionEndSupported';
 import getTransformStyles from '../utils/getTransformStyles';
 
-import getCalendarMonthWidth from '../utils/getCalendarMonthWidth';
+import getCalendarMonthGridWidthStyle from '../utils/getCalendarMonthGridWidthStyle';
 
 import OrientationShape from '../shapes/OrientationShape';
 
@@ -63,19 +63,6 @@ const defaultProps = {
   // i18n
   monthFormat: 'MMMM YYYY', // english locale
 };
-
-function getCalendarMonthGridWidthStyle(orientation, withWeekNumbers) {
-  const calendarMonthWidth = getCalendarMonthWidth(withWeekNumbers);
-
-  switch (orientation) {
-    case HORIZONTAL_ORIENTATION:
-      return { width: 4 * calendarMonthWidth };
-    case VERTICAL_ORIENTATION:
-      return { width: calendarMonthWidth };
-    default:
-      return {};
-  }
-}
 
 export default class CalendarMonthGrid extends React.Component {
   constructor(props) {
